tests input

+++ test/tools/science_and_technology/divisor/logic/divisor_test.dart
@@ -14,7 +14,7 @@ void main() {
 
     _inputsToExpected.forEach((elem) {
       test('input: ${elem['input']}', () {
-        var _actual = divisors(elem['input']);
+        var _actual = divisors(elem['input'] as int);
         expect(_actual, elem['expectedOutput']);
         var length = elem['expectedOutput'].length;
         for (int i = 0; i < length; i++) {
